fix: create tables in a db if they exist

+++ server/migrations/20221014094734__create_tasks_table.js
@@ -1,25 +1,52 @@
 export const up = (knex) => (
-  knex.schema.createTable('tasks', (table) => {
-    table.increments('id').primary();
-    table.string('name');
-    table.string('description');
-    table
-      .integer('status_id')
-      .references('id')
-      .inTable('statuses')
-      .onDelete('RESTRICT');
-    table
-      .integer('creator_id')
-      .references('id')
-      .inTable('users')
-      .onDelete('RESTRICT');
-    table
-      .integer('executor_id')
-      .references('id')
-      .inTable('users')
-      .onDelete('RESTRICT');
-    table.timestamp('created_at').defaultTo(knex.fn.now());
-  })
+  knex.schema.hasTable('tasks')
+    .then((exists) => {
+      if (!exists) {
+        return false;
+      }
+      return knex.schema.createTable('tasks', (table) => {
+        table.increments('id').primary();
+        table.string('name');
+        table.string('description');
+        table
+          .integer('status_id')
+          .references('id')
+          .inTable('statuses')
+          .onDelete('RESTRICT');
+        table
+          .integer('creator_id')
+          .references('id')
+          .inTable('users')
+          .onDelete('RESTRICT');
+        table
+          .integer('executor_id')
+          .references('id')
+          .inTable('users')
+          .onDelete('RESTRICT');
+        table.timestamp('created_at').defaultTo(knex.fn.now());
+      });
+    })
+  // knex.schema.createTable('tasks', (table) => {
+  //   table.increments('id').primary();
+  //   table.string('name');
+  //   table.string('description');
+  //   table
+  //     .integer('status_id')
+  //     .references('id')
+  //     .inTable('statuses')
+  //     .onDelete('RESTRICT');
+  //   table
+  //     .integer('creator_id')
+  //     .references('id')
+  //     .inTable('users')
+  //     .onDelete('RESTRICT');
+  //   table
+  //     .integer('executor_id')
+  //     .references('id')
+  //     .inTable('users')
+  //     .onDelete('RESTRICT');
+  //   table.timestamp('created_at').defaultTo(knex.fn.now());
+  // })
 );
 
 // export const down = (knex) => knex.schema.dropTable('tasks');
